Se reorganiza para que los modulos sean independientes y que esten en su propio py.
