--- Mineria de Datos/db/app/db build.txt
@@ -1,10 +1,0 @@
-mariadb -u root -p
-
-CREATE DATABASE bdd;
-USE bdd;
-
-CREATE TABLE users (
-    id INT PRIMARY KEY AUTO_INCREMENT,
-    username VARCHAR(50) NOT NULL,
-    password_hash VARCHAR(255) NOT NULL
-);
